+++ Packages/dev.unity2rerun.sdk/Runtime/Components/Manager/RerunManager.Lifecycle.cs
@@ -17,7 +17,7 @@ namespace Unity.RerunSDK.Unity
                 Application.runInBackground = true;
 
             MigrateLegacyOutputPath();
-            _encoder = new ManagedRerunEncoder();
+            _encoder = new ManagedRerunEncoder(_recordingCompression);
         }
 
         private void OnValidate()